+++ app/Providers/AuthServiceProvider.php
@@ -33,13 +33,9 @@ class AuthServiceProvider extends ServiceProvider
 
     /**
      * Register any authentication / authorization services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
-        $this->registerPolicies();
-
         // Check permissions of users roles
         Gate::before(function ($user, $ability) {
             if ($user->hasPermission($ability)) {