+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Scheduler not running in production environment ([#323],[#327])
-
+- Database upgrade command not running in production environment ([#324],[#326])
 
 ## [v2.0.0-RC.1] - 2023-03-09
 ### Added
@@ -462,6 +462,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#317]: https://github.com/THM-Health/PILOS/pull/317
 [#316]: https://github.com/THM-Health/PILOS/pull/316
 [#323]: https://github.com/THM-Health/PILOS/issues/323
+[#324]: https://github.com/THM-Health/PILOS/issues/324
+[#326]: https://github.com/THM-Health/PILOS/pull/326
 [#327]: https://github.com/THM-Health/PILOS/pull/327
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v2.0.0-RC.1...HEAD