+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Scheduler not running in production environment ([#323],[#327])
 - Database upgrade command not running in production environment ([#324],[#326])
 
+### Changed
+- Upgrade to Laravel 10 ([#325])
+
 ## [v2.0.0-RC.1] - 2023-03-09
 ### Added
 - Support to customize trusted proxies ([#305],[#306])